Add tool specific lint/fix targets
This commit breaks out the lint/fix commands to their own targets that make
can run concurrently (in parallel or not too!). The benefit of this is that
we can run `make --keep-going` and all the targets will be run regardless
of exit status (while still getting an error exit if any of them errored).
This gives us the benefit of seeing *all* lint issues in one go instead of
stopping at the first errored linter.

This Makefile is also just plain nicer in my opinion. Everything to do with
per file linter is in just one chunk in the file instead of multiple. We can
easily see all the per-$file code. This also means we only have to mess within
the file-type's block to add/remove linters instead of in multiple locations
(block, _lint, and fix).

Signed-off-by: Manuel Mendez <mmendez@equinix.com>