Always use find -execdir for golangci-lint runs
There's not an overwhelming reason to avoid find for the single go module
setup, it works as is for that case. By doing this it means that users won't
have to remember to re-run lint-install if they go from 1 to >1 modules,
which I doubt will happen if we keep them separate. Also, the less corner
cases the better!

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>